Day 18 - LavaductLagoon - part 2 - working sweeping line algorithm - Pick's and Shoelace theorems

+++ src/main/java/advent/day18/Point.java
@@ -3,7 +3,7 @@ package advent.day18;
 import java.util.Objects;
 
 public class Point {
-    int x, y;
+    long x, y;
 
     public Point(int x, int y) {
         this.x = x;
